# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Gin swagger",
        "title": "Gin swagger",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/linebot/{lineuserId}/message": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "linebot"
                ],
                "summary": "fetch messages",
                "parameters": [
                    {
                        "type": "string",
                        "description": "lineuserId",
                        "name": "lineuserId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "linebot"
                ],
                "summary": "push message",
                "parameters": [
                    {
                        "type": "string",
                        "description": "lineuserId",
                        "name": "lineuserId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "msg",
                        "name": "text",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Text"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Text": {
            "type": "object",
            "properties": {
                "text": {
                    "type": "string"
                }
            }
        }
    }
}